[feat] 컬렉션뷰 세그 구현
근데 엉망진창을 곁들인
