+++ Find-iOS/Find-iOS/Screen/ValueTest/ViewController/ValueListVC.swift
@@ -7,13 +7,13 @@
 
 import UIKit
 
+enum Category {
+    case relationship, family, career
+}
+
 class ValueListVC: UIViewController {
     
     //MARK: - Custom Variables
-    
-    enum Category {
-        case relationship, family, career
-    }
     
     var currentCategory: Category = .relationship
     
@@ -32,8 +32,7 @@ class ValueListVC: UIViewController {
     @IBOutlet var segueFamilyBtn: UIButton!
     @IBOutlet var segueCareerBtn: UIButton!
     
-    
-    @IBOutlet var valueListTableView: UITableView!
+    @IBOutlet var valueListCollectionView: UICollectionView!
     
     
     //MARK: - Lifecycle Methods
@@ -41,12 +40,12 @@ class ValueListVC: UIViewController {
     override func viewDidLoad() {
         super.viewDidLoad()
         
-        valueListTableView.dataSource = self
-        
         setSegueStyle()
         RelationshipSelected()
         finishBtnDeactivated()
         
+        valueListCollectionView.dataSource = self
+        valueListCollectionView.delegate = self
     }
     
     
@@ -62,63 +61,20 @@ class ValueListVC: UIViewController {
     
     @IBAction func relationshipBtnDidTap(_ sender: Any) {
         RelationshipSelected()
-        valueListTableView.reloadData()
+        valueListCollectionView.reloadData()
     }
     
     @IBAction func familyBtnDidTap(_ sender: Any) {
         FamilySelected()
-        valueListTableView.reloadData()
+        valueListCollectionView.reloadData()
     }
     
     @IBAction func careerBtnDidTap(_ sender: Any) {
         CareerSelected()
-        valueListTableView.reloadData()
+        valueListCollectionView.reloadData()
     }
     
-    @objc func selectBtnDidTap(_ sender: UIButton) {
-        print("tap \(sender.tag)")
-        
-        var currIndexPath = 0
-        
-        switch self.currentCategory {
-        case .relationship:
-            currIndexPath = sender.tag
-        case .family:
-            currIndexPath = sender.tag + 10
-        case .career:
-            currIndexPath = sender.tag + 20
-        }
-        
-        if valueQuestions[currIndexPath].userChoice == 0 { /// 답변이 안 된 질문이면
-            self.showToast(message: "답변을 하지 않아 선택이 불가합니다")
-        } else { /// 답변을 한 문항이면
-            if selectedCount >= 5 { /// 이미 5개가 선택된 상황이면
-                if valueQuestions[currIndexPath].isChosen { /// 선택 했던 걸 선택하면
-                    /// 선택 취소
-                    valueQuestions[currIndexPath].isChosen = !valueQuestions[currIndexPath].isChosen
-                    self.selectedCount -= 1
-                    self.chosenQuestions.remove(at: self.chosenQuestions.firstIndex(of: valueQuestions[currIndexPath].id)!)
-                } else { /// 새로운 걸 선택하면
-                    self.showToast(message: "최대 5개까지 선택이 가능합니다")
-                }
-            } else { /// 아직 5개가 안 됐으면
-                valueQuestions[currIndexPath].isChosen = !valueQuestions[currIndexPath].isChosen
-                
-                if valueQuestions[currIndexPath].isChosen { /// 새로 선택된 거면
-                    self.selectedCount += 1
-                    self.chosenQuestions.append(valueQuestions[currIndexPath].id)
-                    print(self.chosenQuestions)
-                } else { /// 선택 해제된 거면
-                    self.selectedCount -= 1
-                    self.chosenQuestions.remove(at: self.chosenQuestions.firstIndex(of: valueQuestions[currIndexPath].id)!)
-                }
-            }
-        }
-        
-        
-        
-        valueListTableView.reloadData()
-    }
+    
 }
 
 extension ValueListVC {
@@ -201,114 +157,30 @@ extension ValueListVC {
         segueFamilyBtn.setTitleColor(.subGray2, for: .normal)
     }
     
-    /// 답변이 있는 질문 답변 뷰 생성
-    func answerFilled(cell: ValueListTVC, indexPath: Int) {
-        cell.answerView.makeRounded(cornerRadius: 16.5)
-        cell.answerView.backgroundColor = .white
-        cell.answerView.layer.borderWidth = 1
-        cell.answerView.layer.borderColor = UIColor.find_DarkPurple.cgColor
-        
-        cell.answerLabel.textColor = .find_DarkPurple
-        cell.answerLabel.text = valueQuestions[indexPath].choice[valueQuestions[indexPath].userChoice - 1].choiceContent
-    }
     
-    /// 답변이 없는 질문  답변 뷰 생성
-    func answerUnfilled(cell: ValueListTVC) {
-        cell.answerView.makeRounded(cornerRadius: 16.5)
-        cell.answerView.backgroundColor = .white
-        cell.answerView.layer.borderWidth = 1
-        cell.answerView.layer.borderColor = UIColor.subGray2.cgColor
-        
-        cell.answerLabel.text = "아직 답변을 안 했어요!"
-        cell.answerLabel.textColor = .subGray2
-        cell.answerLabel.font = .spoqaRegular(size: 12)
-        
-    }
-    
-    /// 선택된 답변일 때
-    func selectedQuestion(cell: ValueListTVC, currIndexPath: Int) {
-        cell.selectedBoxView.layer.borderWidth = 2
-        cell.selectedBoxView.layer.borderColor = UIColor.find_DarkPurple.cgColor
-        cell.selectedBoxView.backgroundColor = .find_Purple
-        cell.selectedBoxView.makeRounded(cornerRadius: 10)
-        
-        cell.selectedCountLabel.isHidden = false
-        cell.selectedCountLabel.font = .spoqaBold(size: 9)
-        cell.selectedCountLabel.textColor = .subGray6
-        
-        cell.selectedCountLabel.text = "\((chosenQuestions.firstIndex(of: valueQuestions[currIndexPath].id) ?? 0) + 1 )"
-    }
-    
-    /// 선택되지 않은 답변일 때
-    func unselectedQuestion(cell: ValueListTVC) {
-        cell.selectedBoxView.layer.borderWidth = 2
-        cell.selectedBoxView.layer.borderColor = UIColor.subGray1.cgColor
-        cell.selectedBoxView.backgroundColor = .white
-        cell.selectedBoxView.makeRounded(cornerRadius: 10)
-        
-        cell.selectedCountLabel.isHidden = true
-    }
     
     //MARK: - functions
     
-    
 }
 
-extension ValueListVC: UITableViewDataSource {
-    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return 10
+extension ValueListVC: UICollectionViewDataSource {
+    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
+        3
     }
     
-    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        
-        guard let cell = tableView.dequeueReusableCell(withIdentifier: "ValueListTVC") as? ValueListTVC
-        
+    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
+        guard let cell = collectionView.dequeueReusableCell(withReuseIdentifier: "ValueListCVC", for: indexPath) as? ValueListCVC
         else {
-            return UITableViewCell()
+            return UICollectionViewCell()
         }
-        
-        /// 질문 뷰 스타일
-        cell.questionNumberLabel.text = "Q \(indexPath.row + 1)"
-        cell.questionNumberLabel.font = .spoqaRegular(size: 16)
-        cell.questionNumberLabel.textColor = .subGray3
-        
-        cell.questionLabel.font = .spoqaRegular(size: 13)
-        cell.questionLabel.textColor = .subGray3
-        
-        /// 선택 카테고리에 따른 질문
-        var currIndexPath: Int = 0
-        
-        switch currentCategory {
-        case .relationship:
-            currIndexPath = indexPath.row
-            cell.questionLabel.text = valueQuestions[currIndexPath].question
-        case .family:
-            currIndexPath = indexPath.row + 10
-            cell.questionLabel.text = valueQuestions[currIndexPath].question
-        case .career:
-            currIndexPath = indexPath.row + 20
-            cell.questionLabel.text = valueQuestions[currIndexPath].question
-        }
-        
-        /// 답변이 있을 때 / 없을 때 분기처리
-        if valueQuestions[currIndexPath].userChoice != 0 {
-            answerFilled(cell: cell, indexPath: currIndexPath)
-        } else {
-            answerUnfilled(cell: cell)
-        }
-
-        
-        /// 선택이 된 답변 / 안 된 답변 분기처리
-        if valueQuestions[currIndexPath].isChosen {
-            selectedQuestion(cell: cell, currIndexPath: currIndexPath)
-        } else {
-            unselectedQuestion(cell: cell)
-        }
-        
-        cell.selectedBtn.tag = indexPath.row
-        cell.selectedBtn.addTarget(self, action: #selector(selectBtnDidTap), for: .touchUpInside)
         
         return cell
     }
     
+}
+
+extension ValueListVC: UICollectionViewDelegateFlowLayout {
+    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
+        return CGSize(width: view.frame.width, height: view.frame.height)
+    }
 }
